pyproject.toml
[project]
name = "comfyui-whisper"
description = "Transcribe audio and add subtitles to videos using Whisper in ComfyUI"
version = "1.0.0"
license = {file = "LICENSE"}
dependencies = ["openai-whisper", "pillow", "torchaudio"]

[project.urls]
Repository = "https://github.com/yuvraj108c/ComfyUI-Whisper"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "yuvraj108c"
DisplayName = "ComfyUI-Whisper"
Icon = ""
